+++ app/src/main/java/org/tasks/billing/BillingClient.java
@@ -1,6 +1,7 @@
 package org.tasks.billing;
 
 import android.app.Activity;
+import androidx.annotation.Nullable;
 import androidx.lifecycle.LifecycleOwner;
 import androidx.lifecycle.Observer;
 import java.util.List;
@@ -19,5 +20,7 @@ public interface BillingClient {
 
   void consume(String sku);
 
-  void initiatePurchaseFlow(Activity activity, String sku, String skuType);
+  void initiatePurchaseFlow(Activity activity, String sku, String skuType, @Nullable String oldSku);
+
+  void addPurchaseCallback(OnPurchasesUpdated onPurchasesUpdated);
 }