+++ app/src/main/java/org/tasks/billing/OnPurchasesUpdated.java
@@ -1,0 +1,5 @@
+package org.tasks.billing;
+
+public interface OnPurchasesUpdated {
+  void onPurchasesUpdated();
+}