+++ app/src/main/java/org/tasks/dialogs/IconLayoutManager.java
@@ -5,11 +5,11 @@ import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import org.tasks.R;
 
-class IconLayoutManager extends GridLayoutManager {
+public class IconLayoutManager extends GridLayoutManager {
 
   private int iconSize;
 
-  IconLayoutManager(Context context) {
+  public IconLayoutManager(Context context) {
     super(context, DEFAULT_SPAN_COUNT, RecyclerView.VERTICAL, false);
     this.iconSize = (int) context.getResources().getDimension(R.dimen.icon_picker_size);
   }
@@ -17,7 +17,7 @@ class IconLayoutManager extends GridLayoutManager {
   @Override
   public void onLayoutChildren(RecyclerView.Recycler recycler, RecyclerView.State state) {
     int width = getWidth();
-    if (getSpanCount() == DEFAULT_SPAN_COUNT && iconSize > 0 && width > 0 && getHeight() > 0) {
+    if (getSpanCount() == DEFAULT_SPAN_COUNT && width > 0) {
       setSpanCount(Math.max(1, (width - getPaddingRight() - getPaddingLeft()) / iconSize));
     }
     super.onLayoutChildren(recycler, state);