+++ app/src/main/java/org/tasks/preferences/BasicPreferences.java
@@ -34,6 +34,7 @@ import org.tasks.activities.ColorPickerActivity.ColorPalette;
 import org.tasks.analytics.Tracker;
 import org.tasks.analytics.Tracking;
 import org.tasks.analytics.Tracking.Events;
+import org.tasks.billing.BillingClient;
 import org.tasks.billing.Inventory;
 import org.tasks.billing.PurchaseActivity;
 import org.tasks.dialogs.DialogBuilder;
@@ -81,6 +82,7 @@ public class BasicPreferences extends InjectingPreferenceActivity
   @Inject Device device;
   @Inject ActivityPermissionRequestor permissionRequestor;
   @Inject GoogleAccountManager googleAccountManager;
+  @Inject BillingClient billingClient;
 
   private Bundle result;
 
@@ -221,22 +223,21 @@ public class BasicPreferences extends InjectingPreferenceActivity
             });
 
     Preference upgradeToPro = findPreference(R.string.upgrade_to_pro);
+    upgradeToPro.setOnPreferenceClickListener(
+        p -> {
+          startActivity(new Intent(this, PurchaseActivity.class));
+          return false;
+        });
     if (inventory.hasPro()) {
       upgradeToPro.setTitle(R.string.manage_subscription);
-      upgradeToPro.setOnPreferenceClickListener(
-          p -> {
-            startActivity(
-                new Intent(
-                    Intent.ACTION_VIEW, Uri.parse(getString(R.string.manage_subscription_url))));
-            return false;
-          });
-    } else {
-      upgradeToPro.setOnPreferenceClickListener(
-          p -> {
-            startActivity(new Intent(this, PurchaseActivity.class));
-            return false;
-          });
+      upgradeToPro.setSummary(R.string.manage_subscription_summary);
     }
+
+    findPreference(R.string.refresh_purchases).setOnPreferenceClickListener(
+        preference -> {
+          billingClient.queryPurchases();
+          return false;
+        });
 
     findPreference(R.string.changelog)
         .setSummary(getString(R.string.version_string, BuildConfig.VERSION_NAME));
@@ -251,7 +252,12 @@ public class BasicPreferences extends InjectingPreferenceActivity
 
     //noinspection ConstantConditions
     if (BuildConfig.FLAVOR.equals("generic")) {
-      requires(R.string.about, false, R.string.rate_tasks, R.string.upgrade_to_pro);
+      requires(
+          R.string.about,
+          false,
+          R.string.rate_tasks,
+          R.string.upgrade_to_pro,
+          R.string.refresh_purchases);
       requires(R.string.privacy, false, R.string.p_collect_statistics);
     }